add filer aggregation to class type filter
